refactor(app): server/generation 包更名为 orchestrator(生成任务编排/调度)
该包职责是**生成任务的编排调度**(建 PENDING 记录 → 后台 executor 驱动 → 查询状态),
实际 AI 生成在 ai_engine;原名 generation 易与"生成本身"混淆,更名为 orchestrator。

- git mv server/generation → server/orchestrator(rename 保留历史)
- 全部 windup_app.server.generation → .orchestrator(bootstrap / web api / 包内自引用 / 测试)
- 分层契约不变(约束在 windup_app.server 层级,不牵子包名)

门禁:ruff / lint-imports(2 kept) / 编排测试(3 passed) 全绿。类名(GenerationTask 等)
与 web 路由不动(仍描述生成任务,改动最小)。
